feat: Workspace refactor and tutorials
