+++ Cargo.toml
@@ -1,48 +1,67 @@
-[package]
-name = "bevy_rand"
-version = "0.4.0"
-edition = "2021"
+[workspace]
+members = ["bevy_prng"]
+resolver = "2"
+
+[workspace.package]
 authors = ["Gonçalo Rica Pais da Silva <bluefinger@gmail.com>"]
-description = "A plugin to integrate rand for ECS optimised RNG for the Bevy game engine."
+edition = "2021"
 repository = "https://github.com/Bluefinger/bevy_rand"
 license = "MIT OR Apache-2.0"
+version = "0.5.0"
+rust-version = "1.76.0"
+
+[workspace.dependencies]
+bevy = { version = "0.13", default-features = false }
+serde = "1"
+serde_derive = "1"
+rand_core = { version = "0.6", features = ["std"] }
+rand_chacha = "0.3"
+wyrand = "0.1"
+rand_pcg = "0.3"
+rand_xoshiro = "0.6"
+
+[package]
+name = "bevy_rand"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+description = "A plugin to integrate rand for ECS optimised RNG for the Bevy game engine."
+repository = { workspace = true }
+license = { workspace = true }
 keywords = ["game", "bevy", "rand", "rng"]
 categories = ["game-engines", "algorithms"]
 exclude = ["/.*"]
-resolver = "2"
-rust-version = "1.70.0"
+rust-version = { workspace = true }
 
 [features]
 default = ["serialize", "thread_local_entropy"]
 thread_local_entropy = ["dep:rand_chacha"]
-serialize = ["dep:serde", "rand_core/serde1"]
+serialize = ["dep:serde", "dep:serde_derive", "rand_core/serde1"]
 rand_chacha = ["bevy_prng/rand_chacha"]
 rand_pcg = ["bevy_prng/rand_pcg"]
 rand_xoshiro = ["bevy_prng/rand_xoshiro"]
 wyrand = ["bevy_prng/wyrand"]
 
-[workspace]
-members = ["bevy_prng"]
-
 [dependencies]
 # bevy
-bevy = { version = "0.12.0", default-features = false }
-bevy_prng = { path = "bevy_prng", version = "0.2" }
+bevy.workspace = true
+bevy_prng = { path = "bevy_prng", version = "0.5" }
 
 # others
-serde = { version = "1.0", features = ["derive"], optional = true }
-rand_core = { version = "0.6", features = ["std"] }
-rand_chacha = { version = "0.3", optional = true }
+rand_core.workspace = true
+rand_chacha = { workspace = true, optional = true }
+serde = { workspace = true, optional = true }
+serde_derive = { workspace = true, optional = true }
 
 # This cfg cannot be enabled, but it forces Cargo to keep bevy_prng's
 # version in lockstep with bevy_rand, so that even minor versions
 # cannot be out of step with bevy_rand due to dependencies on traits
 # and implementations between the two crates.
 [target.'cfg(any())'.dependencies]
-bevy_prng = { path = "bevy_prng", version = "=0.2" }
+bevy_prng = { path = "bevy_prng", version = "=0.5" }
 
 [dev-dependencies]
-bevy_prng = { path = "bevy_prng", version = "0.2", features = ["rand_chacha", "wyrand"] }
+bevy_prng = { path = "bevy_prng", version = "0.5", features = ["rand_chacha", "wyrand"] }
 rand = "0.8"
 ron = { version = "0.8.0", features = ["integer128"] }
