+++ src/plugin.rs
@@ -11,19 +11,18 @@ use rand_core::SeedableRng;
 /// use bevy::prelude::*;
 /// use bevy_rand::prelude::*;
 /// use rand_core::RngCore;
-/// use bevy_prng::{ChaCha8Rng, ChaCha12Rng};
 ///
 /// fn main() {
 ///  App::new()
 ///    .add_plugins((
 ///        EntropyPlugin::<ChaCha8Rng>::default(),
-///        EntropyPlugin::<ChaCha12Rng>::default()
+///        EntropyPlugin::<WyRand>::default()
 ///    ))
 ///    .add_systems(Update, print_random_value)
 ///    .run();
 /// }
 ///
-/// fn print_random_value(mut rng: ResMut<GlobalEntropy<ChaCha8Rng>>) {
+/// fn print_random_value(mut rng: ResMut<GlobalEntropy<WyRand>>) {
 ///   println!("Random value: {}", rng.next_u32());
 /// }
 /// ```
