+++ src/tutorial.rs
@@ -1,0 +1,22 @@
+//! # `bevy_rand` Tutorial
+//! 
+//! This is a tutorial covering the various parts of how to use the crate effectively,
+//! from simple usage to more advanced patterns and concepts.
+//! 
+//! Start at [chapter 0](chapter00_overview) if you want to know more why `bevy_rand` exists,
+//! else if you just want to learn how to use this crate, start at [chapter 1](chapter01_choosing_prng)
+//! instead.
+//! 
+//! Note: The content provided is not stable and may change with new versions.
+
+#[doc = include_str!("../tutorial/chapter00-overview.md")]
+pub mod chapter00_overview {}
+
+#[doc = include_str!("../tutorial/chapter01-choosing-prng.md")]
+pub mod chapter01_choosing_prng {}
+
+#[doc = include_str!("../tutorial/chapter02-basic-usage.md")]
+pub mod chapter02_basic_usage {}
+
+#[doc = include_str!("../tutorial/chapter03-components-forking.md")]
+pub mod chapter03_components_forking {}
