+++ tests/determinism.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::type_complexity)]
 
 use bevy::prelude::*;
-use bevy_prng::{ChaCha8Rng, WyRand, ChaCha12Rng};
+use bevy_prng::{ChaCha12Rng, ChaCha8Rng, WyRand};
 use bevy_rand::prelude::*;
 use rand::prelude::Rng;
